[modify] Model - separate the viewModels and Entity, modify the Index name to QueryInit in Service
- 區分viewModels跟Entity
- 將Service層的Index改成QueryInit
