+++ t02_ShopCMS/t02_ShopCMS/Controllers/ProductsController.cs
@@ -11,6 +11,7 @@ using t02_ShopCMS.Models;
 using t02_ShopCMS.Services;
 using Microsoft.Extensions.DependencyInjection;
 using System.Collections.Generic;
+using t02_ShopCMS.Entity;
 
 namespace t02_ShopCMS.Controllers
 {
@@ -28,7 +29,7 @@ namespace t02_ShopCMS.Controllers
         // GET: Products
         public async Task<IActionResult> Index(string searchString)
         {
-            var result = await _productsService.Index(searchString);
+            var result = await _productsService.QueryInit(searchString);
             return View(result);
         }
