+++ t02_ShopCMS/t02_ShopCMS/Entity/ShipmentList.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace t02_ShopCMS.Models
+namespace t02_ShopCMS.Entity
 {
     public class ShipmentList
     {
