+++ t02_ShopCMS/t02_ShopCMS/Models/DetailViewModel.cs
@@ -2,11 +2,12 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using t02_ShopCMS.Entity;
 namespace t02_ShopCMS.Models
 {
     public class DetailViewModel
     {
-        public Models.Product Product { get; set; }
+        public Product Product { get; set; }
         public string Imgsrc { get; set; }
     }
 }
