+++ t02_ShopCMS/t02_ShopCMS/Services/IProductsService.cs
@@ -2,13 +2,14 @@
 using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using t02_ShopCMS.Entity;
 using t02_ShopCMS.Models;
 
 namespace t02_ShopCMS.Services
 {
     public interface IProductsService
     {
-        Task<Indexresp> Index(string searchString);
+        Task<Indexresp> QueryInit(string searchString);
         Task<List<Product>> CategoryFilter(int id);
         Task<bool> Create(Product product, IFormFile myImg);
