+++ t02_ShopCMS/t02_ShopCMS/Views/Products/CreateCategory.cshtml
@@ -1,4 +1,4 @@
-﻿@model Category
+﻿@model t02_ShopCMS.Entity.Category
 @*
     For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 *@
