+++ t02_ShopCMS/t02_ShopCMS/Views/Products/Delete.cshtml
@@ -1,4 +1,4 @@
-﻿@model Product
+﻿@model t02_ShopCMS.Entity.Product
 
 @{
     ViewData["Title"] = "Delete";
