+++ t02_ShopCMS/t02_ShopCMS/Views/Shipment/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<t02_ShopCMS.Models.ShipmentList>
+﻿@model IEnumerable<t02_ShopCMS.Entity.ShipmentList>
 @*
     For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 *@
